micro_ui/requirements.txt
apiflask
flask_cors